order/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
SQLAlchemy==1.4.36
kafka-python==2.0.2
requests==2.28.1
